requirements.txt
tensorflow==2.5.0
pandas==1.2.4
numpy==1.19.5
sklearn==0.0
keras==2.4.3
matplotlib==3.2.2
paho.mqtt==1.5.1
requests==2.25.1
